fix(results): table_browser position u16 not i32 (E0308)
- table_browser/src/lib.rs:167 position: (i+1) as i32 → (i+1) as u16 (Column.position is u16 per pgnative_schema_model::column::Column, E0308 expected u16 found i32)
- fixes cargo check failure on main f598cce (check/5_check.txt:167)

+++ crates/results/table_browser/src/lib.rs
@@ -164,7 +164,7 @@ mod tests {
                 id: *id,
                 owner: Id(0),
                 name: (*n).into(),
-                position: (i + 1) as i32,
+                position: (i + 1) as u16,
                 ty: Id(0),
                 nullability: Nullability::NotNull,
                 has_default: false,
